--- app/src/main/java/com/playtogether_android/app/presentation/ui/message/data/request/ReqEnterRoom.kt
@@ -1,6 +1,0 @@
-package com.playtogether_android.app.presentation.ui.message.data.request
-
-data class ReqEnterRoom(
-    val roomId: Int,
-    val recvId : Int
-)
