--- app/src/main/java/com/playtogether_android/app/presentation/ui/message/data/request/ReqSendMessage.kt
@@ -1,6 +1,0 @@
-package com.playtogether_android.app.presentation.ui.message.data.request
-
-data class ReqSendMessage(
-    val recvId : Int,
-    val content  : String
-)
