+++ app/src/main/java/com/playtogether_android/app/presentation/ui/message/socketData/request/ReqEnterRoom.kt
@@ -1,0 +1,6 @@
+package com.playtogether_android.app.presentation.ui.message.socketData.request
+
+data class ReqEnterRoom(
+    val roomId: Int,
+    val recvId : Int
+)
