+++ app/src/main/java/com/playtogether_android/app/presentation/ui/message/socketData/request/ReqSendMessage.kt
@@ -1,0 +1,6 @@
+package com.playtogether_android.app.presentation.ui.message.socketData.request
+
+data class ReqSendMessage(
+    val recvId : Int,
+    val content  : String
+)
